Move function to get highlight color in service
